Fix skipping test_blog_build_and_clean
This test was being accidentally skipped due to a bad copy/paste on my
part. Oops.

After fixing and running the test, it uncovered a bug in
`copy_static_files` that was also fixed as part of this commit.

+++ nanogen/models.py
@@ -204,6 +204,9 @@ class Blog(object):
         layout_static = os.path.join(self.PATHS['layout'], 'static')
         output_static = os.path.join(self.output_dir, 'static')
 
+        if not os.path.isdir(layout_static):
+            return
+
         if os.path.isdir(os.path.join(output_static)):
             shutil.rmtree(output_static)
